pyproject.toml
[tool.poetry]
name = "edu-modules-project"
version = "0.1.0"
description = ""
authors = ["avdeevaa <tess-avdeeva@seznam.cz>"]
readme = "README.md"
packages = [{include = "edu_modules_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^24.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.3"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
